build: Add default setting config options
Since the application is going to be used on both target and host sides,
it's handy to have a different default settings depending on the
configuration options. We added RII_DEFAULT_NAME and RII_DEFAULT_ADDR
being picked up from environment and used to set default ROM-image file
name and default info-section offset.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>